pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "saytype"
version = "0.1.0"
description = "Voice-to-text menu bar app for macOS. Speak and it types into any focused app."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{name = "Adam Bargh"}]
keywords = ["voice", "dictation", "whisper", "macos", "menu-bar", "accessibility", "hands-free"]
classifiers = [
    "Operating System :: MacOS",
    "Topic :: Adaptive Technologies",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "faster-whisper>=1.0",
    "sounddevice>=0.4",
    "webrtcvad>=2.0.10",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "numpy>=1.24",
]

[project.scripts]
saytype = "saytype.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/saytype"]
